Poor hack to make ntags work with gutentags
This is a very poor hack to make ntags work with gutentags. This hack
was made a few years ago hence there not being a better explanation for
all the changes. But basically gutentags or tagbar was using flags which
ntags didn't support, and this little patch solves that.

you could play with these settings in your .vimrc:
	let g:gutentags_ctags_executable_nim = 'ntags'
	au FileType nim let g:tagbar_ctags_bin = 'ntags'
just make sure ntags is in your $PATH